Reorganized the repo
Goal: To have readme file that describes the project in main folder.
Explore folder will have Basemap .pdf and the field report stuff, in an index file
Copperas-creek folder will have the Mapbox map, and lesson 8 products, along with the video and link to the full-sized layout of the cliff exercise.